Add files via upload
Task: Top Cuisines	Determine the top three most common cuisines in the dataset.								
	Calculate the percentage of restaurants that serve each of the top cuisines.								
Task: City Analysis	Identify the city with the highest number of restaurants in the dataset.							
	Calculate the average rating for restaurants in each city.	 
Task: Online Delivery	Determine the percentage of restaurants that offer online delivery.								
									
	Compare the average ratings of restaurants with and without online delivery								
						
	Determine the city with the highest average rating.							
Task: Price Range Distribution	Create a histogram or bar chart to visualize the distribution of price ranges among the restaurants.										
	Calculate the percentage of restaurants in each price range category.